Changed mixSDL_IntVal and TFBSound_IntVal to intptr_t because they must be able to contain pointer to mixSDL_Buffer
git-svn-id: https://sc2.svn.sourceforge.net/svnroot/sc2/trunk/sc2@1080 8092fc87-c524-0410-9efc-e669fe64eaf9

+++ src/sc2code/libs/sound/sound_chooser.h
@@ -40,7 +40,7 @@
  */
 
 typedef intptr_t TFBSound_Object;
-typedef int TFBSound_IntVal;
+typedef intptr_t TFBSound_IntVal;
 typedef const int TFBSound_SourceProp;
 typedef const int TFBSound_BufferProp;
